requirements.txt
asgiref==3.7.2
black==24.10.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.8
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-crontab==0.7.1
djangorestframework==3.14.0
drf-extensions==0.7.1
feedparser==6.0.11
gunicorn==21.2.0
idna==3.6
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
pytz==2024.1
requests==2.31.0
sgmllib3k==1.0.0
sqlparse==0.4.4
tomli==2.2.1
typing_extensions==4.9.0
urllib3==2.2.1
setuptools==69.1.1
